Extract response generator from ErrorCatcher middleware
