+++ config/di-web.php
@@ -2,6 +2,8 @@
 
 declare(strict_types=1);
 
+use Yiisoft\ErrorHandler\ThrowableHandlerInterface;
+use Yiisoft\ErrorHandler\Handler\ThrowableHandler;
 use Yiisoft\ErrorHandler\Renderer\HtmlRenderer;
 use Yiisoft\ErrorHandler\ThrowableRendererInterface;
 
@@ -10,5 +12,6 @@ use Yiisoft\ErrorHandler\ThrowableRendererInterface;
  */
 
 return [
+    ThrowableHandlerInterface::class => ThrowableHandler::class,
     ThrowableRendererInterface::class => HtmlRenderer::class,
 ];
